Extract the `<VotePreview />` component
